Initialize the Firebase JS APIs
Locally, use a dummy project ID, and manually cofigure firestore
to point to the local emulator.

For production use the Firebase Hosting reserved URL which will
contain values for the project which the hosting is within.

For more information see:
https://firebase.google.com/docs/hosting/reserved-urls

Also add a make command to boot up the firestore emulator and
the application simultaneously.

  $ make start