Move the gateway out of the component props
I dislike having to pass the gateway in as a prop to the component,
so I'm trying this design as an alternative depdendency inversion
technique.

At run-time, I imagine all ItemLists will share the same gateway,
and the props of the ItemList can then be used for rendering different
variants of that component.  Passing in the gateway as a prop
everywhere the ItemList is used feels painful.

Nevertheless I want to keeping the dependency inversion as it allows
me to use a fake gateway for tests and Storybook.